+++ dicomfix/config.py
@@ -44,6 +44,7 @@ class Config:
         self.verbosity = parsed_args.verbosity
         self.range_shifter = self.parse_range_shifter(parsed_args.range_shifter)
         self.repainting = parsed_args.repainting
+        self.repeat_layers = parsed_args.repeat_layers
 
     @staticmethod
     def parse_angles(angles):